+++ Changelog.md
@@ -1,7 +1,9 @@
 # Change Log
 All major changes are listed here.
+# [V0.2.0]
+- [x] Reorganize the code
 
 # [v0.1.0]
 ## Added
-- [] Evaluate the predicted trajectory
-- [] Draw gt vs predicted trajectory
+- [x] Evaluate the predicted trajectory
+- [x] Draw gt vs predicted trajectory